+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## v46 (1/10/2012)
+## v47 (1/30/2013)
+
+Features:
+
+* Disable ruby deploys due to rubygems.org compromise
+
+## v46 (1/10/2013)
 
 Features: